feat: update create model form
