+++ packages/toolkit/src/view/model/CreateModelWithPresetForm/CreateModelWithPresetForm.tsx
@@ -6,7 +6,6 @@ import {
 } from "@instill-ai/design-system";
 import { Nullable, useCreateResourceFormStore } from "../../../lib";
 
-import { CreateModelConfirmationModal } from "./CreateModelConfirmationModal";
 import { SelectModelPresetCard } from "./SelectModelPresetCard";
 import { InputModelBasicInfoCard } from "./InputModelBasicInfoCard";
 
@@ -25,40 +24,16 @@ export const CreateModelWithPresetForm = ({
   width,
   accessToken,
 }: CreateModelWithPresetFormProps) => {
-  const [createModelMessageBoxState, setCreateModelMessageBoxState] =
-    useState<ProgressMessageBoxState>({
-      activate: false,
-      message: null,
-      description: null,
-      status: null,
-    });
-
   return (
     <FormRoot marginBottom={marginBottom} formLess={false} width={width}>
-      <div className="flex flex-col gap-y-6">
-        <div className="flex flex-col">
-          <SelectModelPresetCard marginBottom="mb-10" />
-          <InputModelBasicInfoCard marginBottom={null} />
-          <CreateModelConfirmationModal
-            setErrorMessageBoxState={setCreateModelMessageBoxState}
-            accessToken={accessToken}
-          />
-        </div>
-        <div className="mt-auto">
-          {createModelMessageBoxState.status === "error" ? (
-            <BasicProgressMessageBox
-              state={createModelMessageBoxState}
-              setActivate={(activate) =>
-                setCreateModelMessageBoxState((prev) => ({
-                  ...prev,
-                  activate,
-                }))
-              }
-              width="w-[25vw]"
-              closable={true}
-            />
-          ) : null}
-        </div>
+      <div className="flex flex-col">
+        <SelectModelPresetCard marginBottom="mb-10" />
+        <InputModelBasicInfoCard
+          marginBottom={null}
+          accessToken={accessToken}
+          onCreate={onCreate}
+          initStoreOnCreate={initStoreOnCreate}
+        />
       </div>
     </FormRoot>
   );
